+++ module/Models/Song.cs
@@ -19,7 +19,6 @@ namespace Maestro.Models
         public string CommunityId { get; set; }
         public string BuiltInId { get; set; }
         public bool IsUploaded { get; set; }
-        public bool IsSubmittal { get; set; }
         public int? Bpm { get; set; }
         public SeekData SeekData { get; set; }