+++ module/UI/Main/SongFilterBar.cs
@@ -55,11 +55,7 @@ namespace Maestro.UI.Main
                 IsTextInputFocused = e.Value;
             };
 
-#if DEBUG
-            var sourceItems = new[] { "All", "Favorites", "Bundled", "Created", "Imported", "Community", "Submittals" };
-#else
             var sourceItems = new[] { "All", "Favorites", "Bundled", "Created", "Imported", "Community" };
-#endif
             _filterButton = new GenericFilterButton(
                 new FilterSection { Items = sourceItems, DefaultValue = "All" },
                 new FilterSection { Items = BuildInstrumentFilterItems(), DefaultValue = "All" },